// src/interface/demandados.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"
// To parse this data:
//
//   import { Convert } from "./file";
//
//   const intCarpetaDemandado = Convert.toIntCarpetaDemandado(json);

export interface IntCarpetaDemandado {
  Carpeta: number;
  Demandado: Demandado;
  llaveProceso: string;
  Category: string;
  idProceso?: number[];
  Codeudor?: Codeudor;
}

export interface Codeudor {
  Id: number | string;
  Nombre: string;
  Tel?: Tel;
  Direccion?: string;
}

export interface Tel {
  Celular?: number;
  Fijo?: number;
}

export interface Demandado {
  Id: number;
  Nombre: string;
  Tel: Tel;
  Email?: string;
  Direccion: string;
}

// Converts JSON strings to/from your types
export class Convert {
  public static toIntCarpetaDemandado (
    json: string
  ): IntCarpetaDemandado[] {
    return JSON.parse(
      json
    );
  }

  public static intCarpetaDemandadoToJson (
    value: IntCarpetaDemandado[]
  ): string {
    return JSON.stringify(
      value
    );
  }
}

export interface monCarpetaDemandado extends IntCarpetaDemandado {
  _id: string;
}

export interface intFecha extends IntCarpetaDemandado {
  fecha: Date | string
}